GH-298: [feat] Create separate service to handle kafka message production to orchestration-service.
