+++ Microservices/user-service/src/main/java/app/sportahub/userservice/service/auth/AuthService.java
@@ -1,6 +1,5 @@
 package app.sportahub.userservice.service.auth;
 
-import app.sportahub.kafkevents.ForgotPasswordRequestedEvent;
 import app.sportahub.userservice.dto.request.auth.LoginRequest;
 import app.sportahub.userservice.dto.request.auth.RefreshTokenRequest;
 import app.sportahub.userservice.dto.request.auth.RegistrationRequest;
@@ -19,6 +18,4 @@ public interface AuthService {
     void sendVerificationEmail(String userId);
 
     void sendPasswordResetEmail(String email);
-
-    void sendPasswordResetEmailUsingKafka(String email);
 }
