+++ Microservices/user-service/src/main/java/app/sportahub/userservice/service/kafka/producer/OrchestrationServiceProducer.java
@@ -1,0 +1,7 @@
+package app.sportahub.userservice.service.kafka.producer;
+
+
+public interface OrchestrationServiceProducer {
+
+    void sendPasswordResetEmailUsingKafka(String email);
+}
